requirements.txt
openai>=1.0.0
python-dotenv
gradio
sendgrid>=6.11.0
requests>=2.31.0
